% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_var.R
\name{add_time_if_exists}
\alias{add_time_if_exists}
\title{Helper function to add the time column to the tibble if the time vector is
not NULL}
\usage{
add_time_if_exists(tbl, time)
}
\arguments{
\item{tbl}{The tibble holding the aggregated data}

\item{time}{Time vector}
}
\value{
Returns the tibble with the time columns at the first position if
  time is not NULL
}
\description{
Helper function to add the time column to the tibble if the time vector is
not NULL
}
\keyword{internal}
