% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ncdf.R
\name{aggregate_ncdf}
\alias{aggregate_ncdf}
\title{Aggregate NCDF Climate (Raster) Data for Catchment Subbasins}
\usage{
aggregate_ncdf(ncdf_pth, basin_shp, ncdf_crs, shp_index, var_lbl,
  lat_lbl = "lat", lon_lbl = "lon", time_lbl = "time")
}
\arguments{
\item{ncdf_pth}{Path to the ncdf file}

\item{basin_shp}{Shape file with the basin sub-unit polygons}

\item{ncdf_crs}{Current reference system of ncdf file}

\item{shp_index}{Name of the column in the basin shapefile attribute
table that provides the indices of the basin subunits}
}
\value{
Returns a tibble that provides the time series
  of the aggregated variable for the respective basin subunits
}
\description{
Aggregate NCDF Climate (Raster) Data for Catchment Subbasins
}

