% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_time.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate Time}
\usage{
aggregate_time(ts_tbl, time_int = "day", drop_col = FALSE,
  aggr_fun = mean, ...)
}
\arguments{
\item{ts_tbl}{tibble with timeseries data for basin subunits}

\item{time_int}{character time interval to aggregate, either 'year',
'mon', 'day', 'hour', or 'min'}

\item{drop_col}{logical to set whether unused time columns are dropped
(TRUE) or kept in the table (FALSE). If FALSE unused time columns
values are set to zero.}

\item{aggr_fun}{Aggregation function to apply}

\item{...}{further function arguments of aggr_fun}
}
\value{
Returns an aggregated time series tibble
}
\description{
Aggregate time series derived from climate raster aggregation
}
