% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_var.R
\name{aggregate_variable}
\alias{aggregate_variable}
\title{Intersect var_grid with the shape file and aggregate the variable's data for
the shape file units}
\usage{
aggregate_variable(var_grid, var_data, shape_file, shape_index, time)
}
\arguments{
\item{var_grid}{simple feature polygon. The polygon grid for the individual
pixels of the variable}

\item{var_data}{Tibble where each row represents a pixel in the var_grid and
a column holds the data for one time step in each pixel.}

\item{shape_file}{Shape file with the basin sub-unit polygons.}

\item{shape_index}{Character. Name of the column in the basin shapefile
attribute table that provides the indices of the basin subunits.}

\item{time}{Date Vector for the time steps in var_data}
}
\value{
Returns a tibble that provides the time series of the aggregated
  variable for the respective basin subunits
}
\description{
Intersect var_grid with the shape file and aggregate the variable's data for
the shape file units
}
\keyword{internal}
