% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polygon_grid.R
\name{create_polygon_grid}
\alias{create_polygon_grid}
\title{Create a simple feature object that provides each pixel of the var data as a
polygon, that is used for intersecting with the subunits of the provided
shape file.}
\usage{
create_polygon_grid(var_data, lat_lon, shape_file, crs_ncdf)
}
\arguments{
\item{var_data}{Tibble holding the data of the variable extracted from the
ncdf file}

\item{lat_lon}{List that holds the lat/lon matrices}

\item{shape_file}{The shape file used for later aggregation.}

\item{crs_ncdf}{The reference system of the ncdf file provided as character
string.}
}
\value{
Returns a tibble holding the pixels of the data set as simple feature
  polygons.
}
\description{
Create a simple feature object that provides each pixel of the var data as a
polygon, that is used for intersecting with the subunits of the provided
shape file.
}
\keyword{internal}
