% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polygon_grid.R
\name{extract_poly_coord}
\alias{extract_poly_coord}
\title{Helper function to calculate the coordinates of the pixel polygons.}
\usage{
extract_poly_coord(ind, lat, lon)
}
\arguments{
\item{ind}{Index position of respective pixel.}

\item{lat}{Latitude matrix}

\item{lon}{Longitude matrix}
}
\value{
Returns a list of tables, where each table holds the coordinates
  that describe the path around a cell of the polygon grid.
}
\description{
Helper function to calculate the coordinates of the pixel polygons.
}
\keyword{internal}
