% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polygon_grid.R
\name{extrapol_col}
\alias{extrapol_col}
\title{Helper function to extrapolate the corner points of the grid in longitude.}
\usage{
extrapol_col(mtr)
}
\arguments{
\item{mtr}{matrix for which the longitude boundaries are extrapolated.}
}
\value{
Returns a matrix holding with extrapolated lon boundaries
}
\description{
Helper function to extrapolate the corner points of the grid in longitude.
}
\keyword{internal}
