% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polygon_grid.R
\name{extrapol_row}
\alias{extrapol_row}
\title{Helper function to extrapolate the corner points of the grid in latitude.}
\usage{
extrapol_row(mtr)
}
\arguments{
\item{mtr}{matrix for which the latitude boundaries are extrapolated.}
}
\value{
Returns a matrix holding with extrapolated lat boundaries
}
\description{
Helper function to extrapolate the corner points of the grid in latitude.
}
\keyword{internal}
