% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variables.R
\name{fetch_latlon}
\alias{fetch_latlon}
\title{Fetch the laitude and longitude variables and convert to matrices}
\usage{
fetch_latlon(nc_file, latlon_name)
}
\arguments{
\item{nc_file}{Opened ncdf file}

\item{latlon_name}{Character vector of length 2. Names of latitude longitude
variables (if provided)}
}
\value{
Returns a list holding latitude and longitude matrices
}
\description{
Fetch the laitude and longitude variables and convert to matrices
}
\keyword{internal}
