% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variables.R
\name{fetch_time}
\alias{fetch_time}
\title{Fetch the date vector from the ncdf file if provided}
\usage{
fetch_time(nc_file)
}
\arguments{
\item{nc_file}{Opened ncdf file}
}
\value{
Returns a vector of length 2 givong the names of the lat/lon variables
}
\description{
Fetch the date vector from the ncdf file if provided
}
\keyword{internal}
