% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variables.R
\name{fetch_var}
\alias{fetch_var}
\title{Fetch the trimmed data array according to the trimmed lat/lon and the trimmed
time interval}
\usage{
fetch_var(nc_file, var_name, lat_lon_ind, time_ind)
}
\arguments{
\item{nc_file}{Opened ncdf file}

\item{var_name}{The name of the variable that should be read. If NULL the
first variable is selected}

\item{lat_lon_ind}{The indices of the in x/y dimensions to extract only the
extent of the data array that covers the shape extent}

\item{time_ind}{The indices in the time dimension to only extract the data
for the time period of interest}
}
\value{
Returns the variable's data as a list of matrices.
}
\description{
Fetch the trimmed data array according to the trimmed lat/lon and the trimmed
time interval
}
\keyword{internal}
