% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variables.R
\name{find_latlon}
\alias{find_latlon}
\title{Automatically find latitude and longitude variables}
\usage{
find_latlon(nc_file)
}
\arguments{
\item{nc_file}{Opened ncdf file}
}
\value{
Returns a vector of length 2 givong the names of the lat/lon variables
}
\description{
Automatically find latitude and longitude variables
}
\keyword{internal}
