% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_extent.R
\name{limit_lat}
\alias{limit_lat}
\title{Helper function to find the indices of the latitude matrix that covers
c(xmin,xmax) of the shape file extent}
\usage{
limit_lat(lat, bbox)
}
\arguments{
\item{bbox}{Boundary box of the shape file}

\item{lon}{Matrix holding the latitude values of each pixel}
}
\value{
Returns the updated latitude indices to which the matrices will be
  reduced.
}
\description{
Helper function to find the indices of the latitude matrix that covers
c(xmin,xmax) of the shape file extent
}
\keyword{internal}
