% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_extent.R
\name{limit_lon}
\alias{limit_lon}
\title{Helper function to find the indices of the longitude matrix that covers
c(xmin,xmax) of the shape file extent}
\usage{
limit_lon(lon, bbox)
}
\arguments{
\item{lon}{Matrix holding the longitude values of each pixel}

\item{bbox}{Boundary box of the shape file}
}
\value{
Returns the updated longitude indices to which the matrices will be
  reduced.
}
\description{
Helper function to find the indices of the longitude matrix that covers
c(xmin,xmax) of the shape file extent
}
\keyword{internal}
