% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_var.R
\name{re_sort}
\alias{re_sort}
\title{Helper function to resort the table for its initial index order}
\usage{
re_sort(tbl, grid_int, shp_ind)
}
\arguments{
\item{tbl}{The data table that should be transposed}

\item{grid_int}{sf object from intersection}

\item{shp_ind}{label of shape index}
}
\value{
Returns the transposed tibble with the column names indicated in the
  column name_col
}
\description{
Helper function to resort the table for its initial index order
}
\keyword{internal}
