% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_variables.R
\name{rotate_cc}
\alias{rotate_cc}
\title{Rotate matrix 90° counter clockwise}
\usage{
rotate_cc(mtr)
}
\arguments{
\item{mtr}{Matrix to rotate}
}
\value{
Returns rotated matrix
}
\description{
Rotate matrix 90° counter clockwise
}
\keyword{internal}
