% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_var.R
\name{transpose_tbl}
\alias{transpose_tbl}
\title{Helper function to transpose the data tibble and to name the columns by the
character values in the indicated column}
\usage{
transpose_tbl(tbl, name_col)
}
\arguments{
\item{tbl}{The data table that should be transposed}

\item{name_col}{Character string that defines the name column}
}
\value{
Returns the transposed tibble with the column names indicated in the
  column name_col
}
\description{
Helper function to transpose the data tibble and to name the columns by the
character values in the indicated column
}
\keyword{internal}
