% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_extent.R
\name{trim_time}
\alias{trim_time}
\title{Trim the time vector provided by the ncdf file according to the provided
time_range.}
\usage{
trim_time(time, time_range)
}
\arguments{
\item{time}{Date vector derived from the ncdf file}

\item{time_range}{Vector of length 2 that either provides the date range as
character strings or in a date format. If NULL the complete time vector is
used.}
}
\value{
Returns a list that holds the trimmed date vector and the
  corresponding indices in the original vector that are further used for
  trimming the data set
}
\description{
Trim the time vector provided by the ncdf file according to the provided
time_range.
}
\keyword{internal}
