% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_SWATweather.R
\name{write_SWATweather}
\alias{write_SWATweather}
\title{Write SWAT weather input from aggregated climate variable time series}
\usage{
write_SWATweather(pcp_tbl = NULL, tmp_tbl = NULL, tmn_tbl = NULL,
  tmx_tbl = NULL, hmd_tbl = NULL, slr_tbl = NULL, wnd_tbl = NULL,
  basin_shp, out_type = "ArcSWAT", write_pth)
}
\arguments{
\item{pcp_tbl}{data.frame providing time series of daily precipitation
for the basins subunits}

\item{tmn_tbl}{data.frame providing time series of daily minimum
temperature for the basins subunits}

\item{tmx_tbl}{data.frame providing time series of daily minimum
temperature for the basins subunits}

\item{hmd_tbl}{data.frame providing time series of daily relative
humidity for the basins subunits}

\item{slr_tbl}{data.frame providing time series of daily solar radiation
for the basins subunits}

\item{basin_shp}{Shape file of the basin subunits}

\item{out_type}{Type of output: either "ArcSWAT" to create input files
for ArcSWAT model setup or "txtIO" for direct input into the txtInOut
folder of an existing project (be careful with this method! Might
require changes in file.cio and *.sub files!)}

\item{write_pth}{Path where the SWAT input files are written}

\item{wnd_tbldata.frame}{providing time series of daily average wind
speed for the basins subunits}
}
\value{
Writes Weather input files for a SWAT2012 model
}
\description{
Write SWAT weather input from aggregated climate variable time series
}

