% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ncdf.R
\name{aggregate_ncdf}
\alias{aggregate_ncdf}
\title{Aggregate climate raster data for catchment subbasins}
\usage{
aggregate_ncdf(ncdf_pth, basin_shp, ncdf_crs, shp_index, var_lbl,
  lat_lbl = "lat", lon_lbl = "lon", time_lbl = "time")
}
\arguments{
\item{shp_index}{}
}
\description{
Aggregate climate raster data for catchment subbasins
}

