% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ncdf.R
\name{aggregate_ncdf}
\alias{aggregate_ncdf}
\title{Aggregate ncdf climate (raster) data for catchment subbasins}
\usage{
aggregate_ncdf(
  ncdf_file,
  crs_ncdf,
  shape_file,
  shape_index,
  var_name = NULL,
  latlon_name = NULL,
  time_range = NULL
)
}
\arguments{
\item{ncdf_file}{Path to the ncdf file}

\item{crs_ncdf}{Character string providing the reference system of ncdf file}

\item{shape_file}{Shape file with the basin sub-unit polygons or path to the
file as character string}

\item{shape_index}{Name of the column in the basin shapefile attribute table
that provides the indices of the basin subunits}

\item{var_name}{(optional) character string. Name of the variable array to be
extracted from the ncdf file. If \code{NULL} the first variable provided in
the ncdf file is selected automatically}

\item{latlon_name}{(optional) character vector of \code{length 2}. Names of
the longitude and latitude variables in the ncdf file. Must be provided if
the variables connot be found automatically}

\item{time_label}{(optional) vector of \code{length 2}. Start and end date of
the time range that should be extracted from the ncdf file. Dates can be
either provided as unambiguous text strings (e.g. \code{'yyyy-mm-dd'} or in
a date format}
}
\value{
Returns a tibble that provides the time series of the aggregated
  variable for the respective basin subunits
}
\description{
Aggregate ncdf climate (raster) data for catchment subbasins
}
