% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_timeseries.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate time series derived from climate raster aggregation}
\usage{
aggregate_time(ts_tbl, time_int = "day", aggr_fun = mean)
}
\arguments{
\item{ts_tbl}{tibble with timeseries data for basin subunits}

\item{time_int}{character time interval to aggregate, either 'year',
'mon', 'day', 'hour', or 'min'}

\item{aggr_fun}{Aggregation function to apply}
}
\value{
Returns an aggregated time series tibble
}
\description{
Aggregate time series derived from climate raster aggregation
}

